pyproject.toml
[project]
name = "research-intelligence"
version = "0.2.0"
description = "Agentic research intelligence pipeline for VC diligence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "fastembed>=0.4.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pymupdf>=1.24.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "papers_mcp", "orchestration", "agents", "tools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
